+++ simulations/geometryRTheta/diocotron/diocotron.cpp
@@ -222,9 +222,9 @@ int main(int argc, char** argv)
             to_physical_mapping,
             advection_operator,
             mesh_rtheta,
-            builder_host,
+            builder,
             poisson_solver,
-            spline_evaluator_extrapol_host);
+            spline_evaluator_extrapol);
 #endif
 
     // ================================================================================================