+++ src/geometryRTheta/time_solver/bsl_predcorr_second_order_implicit.hpp
@@ -99,8 +99,8 @@ private:
             PolarBSplinesRTheta,
             SplineRThetaEvaluatorNullBound> const& m_poisson_solver;
 
-    SplineRThetaBuilder_host const& m_builder;
-    SplineRThetaEvaluatorConstBound_host const& m_evaluator;
+    SplineRThetaBuilder const& m_builder;
+    SplineRThetaEvaluatorConstBound const& m_evaluator;
 
 
 
@@ -130,13 +130,13 @@ public:
             LogicalToPseudoPhysicalMapping const& logical_to_pseudo_physical,
             BslAdvectionRTheta const& advection_solver,
             IdxRangeRTheta const& grid,
-            SplineRThetaBuilder_host const& builder,
+            SplineRThetaBuilder const& builder,
             PolarSplineFEMPoissonLikeSolver<
                     GridR,
                     GridTheta,
                     PolarBSplinesRTheta,
                     SplineRThetaEvaluatorNullBound> const& poisson_solver,
-            SplineRThetaEvaluatorConstBound_host const& advection_evaluator)
+            SplineRThetaEvaluatorConstBound const& advection_evaluator)
         : m_logical_to_physical(logical_to_physical)
         , m_advection_solver(advection_solver)
         , m_foot_finder(